moved PDFNotes.ts, PDFSetPageNumbers.ts and Awesomizr.js to srl/pdf/ and set the correct paths in pdf.ts
